==> src/02_analysis/04_decomposition.qmd <==
# Decomposition

```{r}
library(tidyverse)
library(sf)
library(eurostat)
```



```{r}
df <- read_csv("../../data/proc/dwa_simple.csv")
df_gini <- read_csv("../../data/proc/gini_simple.csv")
```

## Decomposition

follows Shorrocks 1982

$$
G = \sum_{k=1}^K \frac{cov(Y_k,Y)}{var(Y)} G 
$$
- note that the G on RHS has no subscript
- so the fraction = percentage contribution

## Exploring

```{r}
df %>%
  filter(
    REF_AREA == "EU",
    UNIT_MEASURE == "EUR",
    DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10"),
    TIME_PERIOD == "2011-01-01"
  ) %>%
  reframe(
    cov = cov(HW, TW_NET), #change the type of asset here
    var = var(TW_NET),
    contribution = cov / var
  )
```
is it smaller than share of housing?

```{r}
df %>%
  filter(
    REF_AREA == "EU",
    UNIT_MEASURE == "EUR",
    DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10"),
    TIME_PERIOD == "2011-01-01"
  ) %>%
  summarise(
    share = sum(HW) / sum(TW_NET)
  )
```
Yes it is!

## Contributions over time

```{r}
df_contrib <- df %>%
  filter(
    UNIT_MEASURE == "EUR",
    DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10")
  ) %>%
  group_by(TIME_PERIOD, REF_AREA) %>%
  summarise(
    #only positive sources
    BW_contrib = cov(BW, TW) / var(TW),
    BW_share = sum(BW) / sum(TW),
    FW_contrib = cov(FW, TW) / var(TW),
    FW_share = sum(FW) / sum(TW),
    HW_contrib = cov(HW, TW) / var(TW),
    HW_share = sum(HW) / sum(TW),
    DEP_contrib = cov(DEP, TW) / var(TW),
    DEP_share = sum(DEP) / sum(TW),
    # BW_contrib = cov(BW, TW_NET) / var(TW_NET),
    # FW_contrib = cov(FW, TW_NET) / var(TW_NET),
    # HW_contrib = cov(HW, TW_NET) / var(TW_NET),
    # DEP_contrib = cov(DEP, TW_NET) / var(TW_NET),
    # DEBT_contrib = cov(DEBT, TW_NET) / var(TW_NET),
    # contrib_sum = BW_contrib + FW_contrib + HW_contrib + DEP_contrib + DEBT_contrib
  )
```

## Biggest Contributors

```{r}
df_contrib %>%
  filter(
    TIME_PERIOD == "2023-01-01"
  ) %>%
  arrange(desc(HW_contrib))
```
CY, DE, MT = winners of this bad game

## Plot over Time

```{r}
df_contrib %>%
  filter(
    TIME_PERIOD>= "2011-01-01",
    REF_AREA %in% c("EU", "DE", "CY", "BE", "NL")
    ) %>%
  ggplot(aes(x=TIME_PERIOD, y=HW_contrib, color=REF_AREA)) +
  geom_line() 
  
```
